Fix from_pretrained dtype kwarg: dtype -> torch_dtype
transformers 4.51.x requires torch_dtype for from_pretrained; the
bare dtype kwarg was forwarded to Qwen3ForCausalLM.__init__ which
doesn't accept it. Later versions added dtype as an alias but 4.51.x
(which has DynamicCache.key_cache) does not.

+++ week10/nano_sglang/model.py
@@ -16,7 +16,7 @@ class Model:
         self.vocab_size = self.config.vocab_size
 
         self.model = AutoModelForCausalLM.from_pretrained(
-            model_path, dtype=self.dtype, device_map=device,
+            model_path, torch_dtype=self.dtype, device_map=device,
         )
         self.model.eval()
